+++ Classes/Backend/Module/ModuleBase.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Sys25\RnBase\Backend\Module;
-
-class ModuleBase extends BaseScriptClass
-{
-}