+++ Classes/Backend/Utility/BackendUtility.php
@@ -8,7 +8,7 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
 /***************************************************************
  * Copyright notice
  *
- * (c) 2015-2020 René Nitzsche <rene@system25.de>
+ * (c) 2015-2023 René Nitzsche <rene@system25.de>
  * All rights reserved
  *
  * This script is part of the TYPO3 project. The TYPO3 project is
@@ -92,10 +92,12 @@ class BackendUtility
      * Checks if we are in dispatch mode.
      *
      * @return bool
+     *
+     * @deprecated muss entfernt werden. Es gibt nur noch den DISPATCH-Modus
      */
     public static function isDispatchMode()
     {
-        return '_DISPATCH' === $GLOBALS['MCONF']['script'];
+        return true;
     }
 
     /* *** ************************************ *** *