+++ Classes/Domain/Model/BaseModel.php
@@ -11,7 +11,7 @@ use tx_rnbase_util_TCA;
 /***************************************************************
  *  Copyright notice
  *
- *  (c) 2007-2021 Rene Nitzsche <rene@system25.de>
+ *  (c) 2007-2023 Rene Nitzsche <rene@system25.de>
  *  All rights reserved
  *
  *  This script is part of the TYPO3 project. The TYPO3 project is