+++ Classes/Utility/TYPO3.php
@@ -150,6 +150,16 @@ class TYPO3
         return self::isTYPO3VersionOrHigher(11005000);
     }
 
+    /**
+     * Prüft, ob mindestens TYPO3 Version 12.1 vorhanden ist.
+     *
+     * @return bool
+     */
+    public static function isTYPO121OrHigher()
+    {
+        return self::isTYPO3VersionOrHigher(12001000);
+    }
+
     private static $TYPO3_VERSION = false;
 
     /**