+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 Changelog
 ---------
 
+v1.17.0 (??.05.2023)
+
+* Install possible in TYPO3 12.x but still not working
+* Removed class ModuleRunner as not needed anymore
+* Removed class ModuleBase as not needed anymore
+* fixed some PHP warnings
+* improved pipeline checks
+* backwards compatibility for main subpart calucation in ListView
+
 v1.16.12 (02.06.2023)
 * Fix wrong annotation in Sys25\RnBase\Domain\Model\DataModel::offsetSet()
 
@@ -17,6 +26,9 @@ v1.16.9 (24.04.2023)
 
 v1.16.8 (21.04.2023)
 * Maintenance Release to keep out changes for TYPO3 12
+
+v1.16.7 (18.04.2023)
+* obsolete, do not use
 
 v1.16.6 (27.03.2023)
 * Support constant TYPO3